split out Users and Profile components
